More reliable upgrades from old index
The files are only deleted after we have asked the message store
to sync to disk.

We always look for .idx files and load them if any, except on
clean recovery. This allows resuming the upgrade. When a node
is shut down during an upgrade (normally or not) and then
restarted, it is possible that some messages end up duplicated,
however.